fix bag

+++ controllers/contactsControllers.js
@@ -1,5 +1,5 @@
 // import contactsService from "../services/contactsServices.js";
-import HttpError from "../helpers/HttpError.js";
+import {HttpError} from "../helpers/HttpError.js";
 import { Contact } from '../models/contact.js';
 
 export const getAllContacts = async (req, res, next) => {
@@ -7,7 +7,7 @@ export const getAllContacts = async (req, res, next) => {
         const { _id: owner } = req.user;
         const { page = 1, limit = 20} = req.query;
         const skip = (page - 1) * limit;
-        const result = await Contact.find({ owner}, "", { skip, limit:Number(limit)});
+        const result = await Contact.find({owner}, "", { skip, limit:Number(limit)});
         res.json(result);
     }
     catch (error) {
@@ -19,7 +19,7 @@ export const getOneContact = async (req, res, next) => {
     try {
         const { id } = req.params;
         const { _id: owner } = req.user;
-        const result = await Contact.findOne({_id:id,owner});
+        const result = await Contact.findOne({_id:id}).where("owner").equals(owner);
         // const result = await Contact.findById(id);
         if (!result) {
             throw HttpError(404);
@@ -33,7 +33,8 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => { 
     try {
         const { id } = req.params;
-        const result = await Contact.findByIdAndDelete(id);
+        const { _id: owner } = req.user;
+        const result = await Contact.findByIdAndDelete(id).where("owner").equals(owner);
         if (!result) {
             throw HttpError(404);
         }
@@ -59,7 +60,8 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const result = await Contact.findByIdAndUpdate(id, req.body, {new:true});
+        const { _id: owner } = req.user;
+        const result = await Contact.findByIdAndUpdate(id, req.body, { new: true }).where("owner").equals(owner);
         if (!result) {
             throw HttpError(400, error.message)
         }
@@ -71,7 +73,8 @@ export const updateContact = async (req, res, next) => {
 export const updateFavorite = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+        const { _id: owner } = req.user;
+        const result = await Contact.findByIdAndUpdate(id, req.body, { new: true }).where("owner").equals(owner);
         if (!result) {
             throw HttpError(400, error.message)
         }
